requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiofiles==24.1.0
aiosqlite==0.20.0
pydantic==2.9.2
python-dotenv==1.0.1
flask==3.0.0
google-generativeai==0.3.2
google-cloud-firestore==2.13.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
